% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotland-data.R
\docType{data}
\name{scotland}
\alias{scotland}
\title{Scotland Dataset}
\format{A dataframe of 2182 observations of 11 variables.
\describe{
  \item{country}{description}
  \item{school}{description}
  \item{num_boys}{description}
  \item{num_girls}{description}
  \item{type}{description}
  \item{score}{description}
  \item{num_books}{description}
  \item{sex}{description}
  \item{word_knowl}{description}
  \item{homework}{description}
  \item{pop}{description}
}}
\source{
\href{link}{Link to Source}
}
\usage{
data(scotland)
}
\description{
Description
}
\keyword{datasets}
