% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{sensBounds}
\alias{sensBounds}
\title{Function to get model fits and plotting parameters}
\usage{
sensBounds(formula = score ~ num_books + factor(sex) + word_knowl +
  homework + factor(type) + log(pop), grouplevel = ~school,
  data = italy, condition_max = 1000)
}
\description{
Function to get model fits and plotting parameters
}
